Refactor: extract file sending logic
